third_party/blink/web_tests/external/wpt/css/css-cascade/at-scope-parsing-expected.txt

This is a testharness.js-based test.
[FAIL] @scope () is valid
  assert_equals: expected "@scope {\\n}" but got "@scope () {\\n}"
[FAIL] @scope () to () is valid
  assert_equals: expected "@scope {\\n}" but got "@scope () {\\n}"
[FAIL] @scope (.c <> .d) is not valid
  assert_equals: expected 0 but got 1
[FAIL] @scope (.a, .c <> .d) is not valid
  assert_equals: expected 0 but got 1
[FAIL] @scope (.a <> .b, .c) is not valid
  assert_equals: expected 0 but got 1
[FAIL] @scope (div::before) is not valid
  assert_equals: expected 0 but got 1
[FAIL] @scope (div::after) is not valid
  assert_equals: expected 0 but got 1
[FAIL] @scope (slotted(div)) is not valid
  assert_equals: expected 0 but got 1
[FAIL] @scope (.a) to (div::before) is not valid
  assert_equals: expected 0 but got 1
[FAIL] @scope (> &) to (>>) is not valid
  assert_equals: expected 0 but got 1
Harness: the test ran to completion.
